fix: Address review feedback for thread support PR
- Fix handle_channel_read RPC to include thread_parent_id in response
  (thread replies were indistinguishable from top-level messages via CLI)
- Move rpc_channel.rs inline tests to separate rpc_channel_tests.rs file
- Move web.rs inline tests to separate web_tests.rs file
- Add test_channel_read_includes_thread_parent_id to verify RPC fix
- Add HTTP-level test for thread_parent_id filter in history API
- Update test_web_api_channel_history to assert on new id field
- Fix stale comment in test_backward_compatibility_struct_literal
